Fix python code after python 3.8 deprecation of 'time.clock()'
Currently when running the python example with python 3.8 the following error is printed:
> AttributeError: module 'time' has no attribute 'clock'
The reason for this is that in python 3.8 'time.clock()' was deprecated.
> https://bugs.python.org/issue14309
This can be fixed by replacing 'time.clock()' with 'time.perf_counter()' which is essentially the same.
> https://docs.python.org/3.7/library/time.html#time.clock
> https://docs.python.org/3.8/library/time.html#time.perf_counter
To fix this the following shel command was executed in the directory 'Python':
> find . -type f -exec sed -i 's/time\.clock()/time\.perf_counter()/g' {} +

The only catch is that 'time.perf_counter() ' was first introduced in 3.3 which means no backwards compatibility to older versions.