Update NikonDataLoader.py
Fixed the loop on images, previously started at index =1, this led to errors when loading partial datasets as the last entry is out of the array's shape

Also, the recon algos all require the geometry object to have the number of voxels and number of detector pixels stored as integers, not as floats. previous version would not run.